Cleanup Verifier 2. Split out pre-encryption verification.
